requirements.txt
cryptography==46.0.4
FIN==2.1.8
natsort==8.4.0
numpy==2.4.2
opencv_python==4.10.0.84
Pillow==9.5.0
Pillow==12.1.0
pypng==0.20220715.0
pyzbar==0.1.9
qrcode==8.2
skimage==0.0
tensorboardX==2.6.2.2
tensorboardX==2.6.4
torch==2.3.1
torchmetrics==1.7.1
torchvision==0.18.1
transformers==4.42.3
viz==0.0.5
